keep .err files when using valgrind

+++ testone.sh
@@ -71,7 +71,10 @@ do if test -x $file
 		  else
 		      echo -e $KNOWNFAIL
 		      \rm -f $file.diff
-		      \rm $file.err
+		      if [ "$VG" == "" ];
+		      then
+			  \rm $file.err
+		      fi
 		  fi
 	      else
 		  echo -e $FAIL
@@ -81,12 +84,18 @@ do if test -x $file
 	  else
 	      echo -e $OK "(variant=$vari)"
 	      rm $file.diff
-	      rm $file.err
+	      if [ "$VG" == "" ];
+	      then
+		  \rm $file.err
+	      fi
 	  fi
       else
 	  echo -e $OK
 	  rm $file.diff
-	  rm $file.err
+	  if [ "$VG" == "" ];
+	  then
+	      \rm $file.err
+	  fi
       fi
    else
        echo "file $file not found"
